+++ demos/manual/src/main/java/is/codion/manual/notes/NotesDemo.java
@@ -120,8 +120,7 @@ public final class NotesDemo {
 			super(Note.TYPE, connectionProvider);
 			// Set the Note.UPDATED value before we perform an update
 			beforeUpdate().addConsumer(notes ->
-							notes.values().forEach(note ->
-											note.put(Note.UPDATED, LocalDateTime.now())));
+							notes.forEach(note -> note.put(Note.UPDATED, LocalDateTime.now())));
 		}
 	}